Add sound to swift tabata. Add ability to have different requires for swift/js. Add better float support for swift.

+++ src/foam/util/Timer.js
@@ -157,7 +157,7 @@ return a! + (1 + s) * (b!-a!)/2;
       },
       swiftCode: function() {/*
 i+=1
-time  += interval * timeWarp
+time  += Int(Float(interval) * timeWarp)
 second = time /    1000 % 60 << 0;
 minute = time /   60000 % 60 << 0;
 hour   = time / 3600000 % 24 << 0;
